ура надеюсь опять ничего не соберётся

+++ prj.lab/stackarr/stackarr.hpp
@@ -25,7 +25,7 @@ public:
 	~StackArr();
 
 private:
-	int32_t size_ = 0;
+	size_t size_ = 0;
 	int32_t elems_ = 0;
 	Complex* data_ = nullptr;
 	Complex* head_ = 0;
